Move 'actions' up one level
