+++ src/actions/entry_set_comment.rs
@@ -3,7 +3,7 @@ use serde::{Serialize, Deserialize};
 use simple_error::{SimpleResult, SimpleError, bail};
 
 use crate::project::h2project::H2Project;
-use crate::project::actions::Action;
+use crate::actions::Action;
 
 #[derive(Serialize, Deserialize, Debug)]
 struct Forward {
