+++ src/project/mod.rs
@@ -1,4 +1,3 @@
-pub mod actions;
 pub mod h2project;
 pub use h2project::H2Project;
